Add files via upload
This is a prototype to group for one group, with the focus on gender > cgpa > school
I created 2 lists of Male and Female, and sort their CGPA on descending order.
First, I will check the ratio of male/female = boy_girl_ratio
If boy_girl_ratio is between 2/3 and 3/2, we can create balanced groups with 3 female or 2 male/ 2 male and 3 female only!
Then, I will match groups of 4 with boy/high cgpa, boy/low cgpa, girl/high cgpa, girl low/cgpa by picking from the sorted groups

If boy_girl_ratio is not in the range of 2/3 to 3/2, if we insist on making female groups and male groups like above, we will have single-gendered groups
Thus, we must concede and let groups of 3-2 and 4-1 to happen. and fortunately, since 50 is divisible by 5, we can put every single student in the big group into small groups of 5

If there is any leftover students after this, we will only append single students to a group of 4 until there is not groups of 4
The rest is regrouped as the last few groups (or the last group) to create 10 full groups

Ofc this needs a check and swapping afterward with school diversity and cgpa in our desired range (10% of mean), but I will do it tomorrow.